moved tempdb to its own folder, login and sign up pages are completed,and have routes associated
