+++ Client/src/db/tempdb.js
@@ -1,10 +1,12 @@
 const brews = [
-    ///provisional use database while we wait for QUERY_BREW is ready
+    ///Cups, Id, Grind , Water, Weight, and Strenght values are all provisional. We should be pulling those from the database.
+
+    //only values stored in the front end are Name, Image, and steps.
     {
         id: 1,
         cups: 2,
         name: "French Press",
-        image: require("../../images/icons/French Press/SVG/French Press.svg").default,
+        image: require("../images/icons/French Press/SVG/French Press.svg").default,
         grind: "coarse",
         water: "50 liters",
         weight: "1 megaton",
@@ -18,7 +20,7 @@ const brews = [
         id: 2,
         cups: 4,
         name: "Hairo",
-        image: require("../../images/icons/Hairo Vaccum/SVG/Hairo Vaccum .svg").default,
+        image: require("../images/icons/Hairo Vaccum/SVG/Hairo Vaccum .svg").default,
         grind: "fine",
         water: "1 liter",
         weight: "420 g",
@@ -32,7 +34,7 @@ const brews = [
         id: 3,
         cups: 1,
         name: "Aero Press",
-        image: require("../../images/icons/AeroPress/SVG/AeroPress.svg").default,
+        image: require("../images/icons/AeroPress/SVG/AeroPress.svg").default,
         grind: "lite",
         water: "2 mililiters",
         weight: "1 terabite",
@@ -46,7 +48,7 @@ const brews = [
         id: 4,
         cups: 1,
         name: "Bialetti",
-        image: require("../../images/icons/Bialetti/SVG/Bialetti.svg").default,
+        image: require("../images/icons/Bialetti/SVG/Bialetti.svg").default,
         grind: "lite",
         water: "2 mililiters",
         weight: "1 terabite",
@@ -61,7 +63,7 @@ const brews = [
         id: 5,
         cups: 1,
         name: "Chemex",
-        image: require("../../images/icons/Chemex/SVG/Chemex.svg").default,
+        image: require("../images/icons/Chemex/SVG/Chemex.svg").default,
         grind: "lite",
         water: "2 mililiters",
         weight: "1 terabite",
@@ -76,7 +78,7 @@ const brews = [
         id: 6,
         cups: 1,
         name: "Neldrip",
-        image: require("../../images/icons/Nel-Drip/SVG/Nel Drip.svg").default,
+        image: require("../images/icons/Nel-Drip/SVG/Nel Drip.svg").default,
         grind: "lite",
         water: "2 mililiters",
         weight: "1 terabite",
@@ -91,7 +93,7 @@ const brews = [
         id: 7,
         cups: 1,
         name: "V60",
-        image: require("../../images/icons/V60/SVG/V60.svg").default,
+        image: require("../images/icons/V60/SVG/V60.svg").default,
         grind: "lite",
         water: "2 mililiters",
         weight: "1 terabite",
@@ -106,7 +108,7 @@ const brews = [
         id: 8,
         cups: 1,
         name: "Kettle",
-        image: require("../../images/icons/Kettle/SVG/Kettle.svg").default,
+        image: require("../images/icons/Kettle/SVG/Kettle.svg").default,
         grind: "lite",
         water: "2 mililiters",
         weight: "1 terabite",
